# docs/api/ACGS_CODE_ANALYSIS_ENGINE_API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: ACGS Code Analysis Engine API
  version: 1.0.0
  description: |
    Comprehensive API for intelligent code analysis, semantic search, and dependency mapping 
    within the ACGS (AI Constitutional Governance System) ecosystem.
    
    ## Features
    - **Semantic Code Search**: Vector-based similarity search using CodeBERT embeddings
    - **Symbol Analysis**: Detailed code symbol extraction and metadata
    - **Dependency Mapping**: Comprehensive dependency graph construction and traversal
    - **Context Integration**: Bidirectional integration with ACGS Context Service
    - **Constitutional Compliance**: Full integration with ACGS governance framework
    
    ## Authentication
    All endpoints require JWT authentication via the ACGS Auth Service (port 8016).
    Include the JWT token in the Authorization header: `Bearer <token>`
    
    ## Performance Targets
    - **Latency**: P99 < 10ms for cached queries
    - **Throughput**: >100 RPS sustained load
    - **Cache Hit Rate**: >85% for repeated queries
    
    ## Constitutional Compliance
    All responses include constitutional validation signature with hash: `cdd01ef066bc6cf2`
    
  contact:
    name: ACGS Development Team
    url: https://github.com/CA-git-com-co/ACGS
    email: dev@acgs.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8007
    description: Local development server
  - url: https://api.acgs.ai/code-analysis
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/search/semantic:
    get:
      tags:
        - Search
      summary: Semantic Code Search
      description: |
        Performs semantic search across the codebase using vector embeddings.
        Returns code symbols ranked by semantic similarity to the query.
      operationId: searchSemantic
      parameters:
        - name: query
          in: query
          required: true
          description: Natural language or code search query
          schema:
            type: string
            minLength: 1
            maxLength: 500
          example: "function that validates user authentication"
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: language
          in: query
          description: Filter by programming language
          schema:
            type: string
            enum: [python, javascript, typescript, yaml, json, sql]
        - name: symbol_type
          in: query
          description: Filter by symbol type
          schema:
            type: string
            enum: [function, class, variable, import, method, property]
        - name: min_score
          in: query
          description: Minimum similarity score (0.0 to 1.0)
          schema:
            type: number
            minimum: 0.0
            maximum: 1.0
            default: 0.5
      responses:
        '200':
          description: Semantic search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticSearchResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/symbols/{symbol_id}:
    get:
      tags:
        - Symbols
      summary: Get Symbol Details
      description: |
        Retrieves detailed information about a specific code symbol,
        including metadata, source code, and related context.
      operationId: getSymbol
      parameters:
        - name: symbol_id
          in: path
          required: true
          description: Unique identifier for the code symbol
          schema:
            type: string
            format: uuid
        - name: include_source
          in: query
          description: Whether to include source code in response
          schema:
            type: boolean
            default: true
        - name: include_context
          in: query
          description: Whether to include related context from Context Service
          schema:
            type: boolean
            default: true
        - name: include_dependencies
          in: query
          description: Whether to include dependency information
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Symbol details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SymbolDetailResponse'
        '404':
          description: Symbol not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/analysis/trigger:
    post:
      tags:
        - Analysis
      summary: Trigger Code Analysis
      description: |
        Triggers analysis of specific files or the entire codebase.
        Useful for manual re-indexing or processing new files.
      operationId: triggerAnalysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisRequest'
            examples:
              single_file:
                summary: Analyze single file
                value:
                  analysis_type: "file"
                  file_paths: ["services/core/auth/auth_service.py"]
                  force_reanalysis: false
              full_scan:
                summary: Full codebase scan
                value:
                  analysis_type: "full_scan"
                  force_reanalysis: true
      responses:
        '202':
          description: Analysis job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
        '400':
          description: Invalid analysis request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/context/enrich:
    post:
      tags:
        - Context Integration
      summary: Enrich Code Analysis with Context
      description: |
        Enriches code analysis results with contextual information from the ACGS Context Service.
        Creates bidirectional links between code symbols and domain contexts.
      operationId: enrichWithContext
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextEnrichmentRequest'
      responses:
        '200':
          description: Context enrichment completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextEnrichmentResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from ACGS Auth Service (port 8016)
  schemas:
...
    CodeSymbol:
      type: object
      properties:
        id:
          type: string
          format: uuid
        file_path:
          type: string
          maxLength: 512
        relative_path:
          type: string
          maxLength: 512
        symbol_name:
          type: string
          maxLength: 256
        symbol_type:
          type: string
          enum: [function, class, variable, import, method, property]
        language:
          type: string
          enum: [python, javascript, typescript, yaml, json, sql, markdown]
        start_line:
          type: integer
          minimum: 1
        end_line:
          type: integer
          minimum: 1
        signature:
          type: string
          nullable: true
        docstring:
          type: string
          nullable: true
        source_code:
          type: string
          nullable: true
        complexity_score:
          type: integer
          minimum: 0
        is_public:
          type: boolean
        is_deprecated:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        constitutional_hash:
          type: string
          example: "cdd01ef066bc6cf2"
      required:
        - id
        - file_path
        - symbol_name
        - symbol_type
        - language
        - start_line
        - end_line
    Dependency:
      type: object
      properties:
        id:
          type: string
          format: uuid
        source_symbol:
          $ref: '#/components/schemas/CodeSymbol'
        target_symbol:
          allOf:
            - $ref: '#/components/schemas/CodeSymbol'
          nullable: true
        dependency_type:
          type: string
          enum: [import, call, inheritance, composition, reference]
        target_name:
          type: string
          nullable: true
          description: For external dependencies
        target_module:
          type: string
          nullable: true
          description: Module name for external dependencies
        is_external:
          type: boolean
        confidence_score:
          type: number
          minimum: 0.0
          maximum: 1.0
        created_at:
          type: string
          format: date-time
      required:
        - id
        - source_symbol
        - dependency_type
        - is_external
        - confidence_score
    ContextLink:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code_symbol_id:
          type: string
          format: uuid
        context_id:
          type: string
          format: uuid
        context_type:
          type: string
          enum: [DomainContext, PolicyContext, ConstitutionalContext, AgentContext]
        relationship_type:
          type: string
          enum: [implements, uses, related_to, validates, enforces]
        confidence_score:
          type: number
          minimum: 0.0
          maximum: 1.0
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - code_symbol_id
        - context_id
        - context_type
        - relationship_type

    # Search Response Schemas
    SemanticSearchResponse:
      type: object
      properties:
        query:
          type: string
        results:
          type: array
          items:
            type: object
            properties:
              symbol:
                $ref: '#/components/schemas/CodeSymbol'
              similarity_score:
                type: number
                minimum: 0.0
                maximum: 1.0
              context_enrichment:
                type: array
                items:
                  $ref: '#/components/schemas/ContextLink'
        total_results:
          type: integer
          minimum: 0
        query_time_ms:
          type: integer
          minimum: 0
        cache_hit:
          type: boolean
        constitutional_signature:
          type: string
          description: Constitutional compliance signature
      required:
        - query
        - results
        - total_results
        - query_time_ms
        - constitutional_signature
...
    SymbolDetailResponse:
      type: object
      properties:
        symbol:
          $ref: '#/components/schemas/CodeSymbol'
        dependencies:
          type: array
          items:
            $ref: '#/components/schemas/Dependency'
        context_links:
          type: array
          items:
            $ref: '#/components/schemas/ContextLink'
        related_symbols:
          type: array
          items:
            $ref: '#/components/schemas/CodeSymbol'
        file_metadata:
          type: object
          properties:
            file_size:
              type: integer
            last_modified:
              type: string
              format: date-time
            language:
              type: string
            symbol_count:
              type: integer
        constitutional_signature:
          type: string
      required:
        - symbol
        - constitutional_signature

    # Request Schemas
    AnalysisRequest:
      type: object
      properties:
        analysis_type:
          type: string
          enum: [file, directory, full_scan]
        file_paths:
          type: array
          items:
            type: string
            maxLength: 512
        force_reanalysis:
          type: boolean
          default: false
        include_embeddings:
          type: boolean
          default: true
        include_dependencies:
          type: boolean
          default: true
      required:
        - analysis_type
    ContextEnrichmentRequest:
      type: object
      properties:
        symbol_ids:
          type: array
          items:
            type: string
            format: uuid
        context_types:
          type: array
          items:
            type: string
            enum: [DomainContext, PolicyContext, ConstitutionalContext, AgentContext]
        min_confidence:
          type: number
          minimum: 0.0
          maximum: 1.0
          default: 0.7
      required:
        - symbol_ids

    # Response Schemas
    AnalysisJobResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        created_at:
          type: string
          format: date-time
        estimated_completion:
          type: string
          format: date-time
          nullable: true
        constitutional_signature:
          type: string
      required:
        - job_id
        - status
        - created_at
        - constitutional_signature
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
        constitutional_hash:
          type: string
          example: "cdd01ef066bc6cf2"
      required:
        - error
        - message
        - timestamp
        - constitutional_hash