# services/core/code-analysis/code_analysis_service/app/api/v1/router.py
"""
ACGS Code Analysis Engine - API Router
FastAPI router implementing all endpoints from OpenAPI specification.

Constitutional Hash: cdd01ef066bc6cf2
"""

import time
import uuid
from typing import List, Optional, Dict, Any
from datetime import datetime, timezone

from fastapi import APIRouter, Depends, HTTPException, Request, status, Query
from fastapi.responses import JSONResponse

from ...models.schemas import (
    SemanticSearchRequest,
    SemanticSearchResponse,
    AnalysisRequest,
    AnalysisResponse,
    ContextEnrichmentRequest,
    ContextEnrichmentResponse,
    CodeSymbol,
    HealthCheck,
    ErrorResponse
)
from ...utils.constitutional import CONSTITUTIONAL_HASH, ensure_constitutional_compliance
from ...utils.logging import get_logger, log_api_request, log_api_response

logger = get_logger("api.v1")

# Create API router
api_router = APIRouter()


@api_router.get("/health", response_model=HealthCheck, tags=["Health"])
async def health_check(request: Request):
    """
    Health check endpoint for service monitoring.
    """
    start_time = time.time()

    try:
        # Basic health check data
        health_data = {
            "status": "healthy",
            "service": "acgs-code-analysis-engine",
            "version": "1.0.0",
            "checks": {
                "api": "healthy",
                "constitutional_compliance": "validated"
            },
            "uptime_seconds": int(time.time()),  # Simplified uptime
            "last_analysis_job": None
        }

        response = HealthCheck(**ensure_constitutional_compliance(health_data))

        # Log health check
        duration_ms = (time.time() - start_time) * 1000
        logger.info(
            "Health check completed",
            extra={
                "status": "healthy",
                "duration_ms": round(duration_ms, 2),
                "constitutional_hash": CONSTITUTIONAL_HASH
            }
        )

        return response

    except Exception as e:
        logger.error(
            f"Health check error: {e}",
            extra={"constitutional_hash": CONSTITUTIONAL_HASH},
            exc_info=True
        )

        # Return unhealthy status
        health_data = {
            "status": "unhealthy",
            "service": "acgs-code-analysis-engine",
            "version": "1.0.0",
            "checks": {
                "api": "error",
                "constitutional_compliance": "validated"
            },
            "uptime_seconds": int(time.time()),
            "last_analysis_job": None
        }

        return HealthCheck(**ensure_constitutional_compliance(health_data))


# Dependency to get current user from request state
def get_current_user(request: Request) -> Optional[Dict[str, Any]]:
    """Get current user from request state (set by auth middleware)."""
    return getattr(request.state, 'user', None)


def get_user_id(request: Request) -> Optional[str]:
    """Get user ID from request state."""
    return getattr(request.state, 'user_id', None)


def get_request_id(request: Request) -> str:
    """Get request ID from request state."""
    return getattr(request.state, 'request_id', str(uuid.uuid4()))


@api_router.get("/search/semantic", response_model=SemanticSearchResponse, tags=["Search"])
async def semantic_search(
    request: Request,
    query: str = Query(..., min_length=1, max_length=500, description="Search query"),
    limit: int = Query(default=10, ge=1, le=100, description="Maximum number of results"),
    min_confidence: float = Query(default=0.7, ge=0.0, le=1.0, description="Minimum confidence score"),
    symbol_types: Optional[str] = Query(default=None, description="Comma-separated symbol types"),
    file_paths: Optional[str] = Query(default=None, description="Comma-separated file paths"),
    include_embeddings: bool = Query(default=False, description="Include embedding vectors"),
    current_user: Optional[Dict[str, Any]] = Depends(get_current_user),
    user_id: Optional[str] = Depends(get_user_id),
    request_id: str = Depends(get_request_id)
):
    """
    Perform semantic search on code symbols.
    
    Uses vector similarity search to find code symbols matching the query.
    """
    start_time = time.time()
    
    # Log API request
    log_api_request(
        method="GET",
        path="/api/v1/search/semantic",
        user_id=user_id,
        request_id=request_id,
        query=query,
        limit=limit
    )
    
    try:
        # Parse optional parameters
        symbol_type_list = None
        if symbol_types:
            symbol_type_list = [s.strip() for s in symbol_types.split(",")]
        
        file_path_list = None
        if file_paths:
            file_path_list = [p.strip() for p in file_paths.split(",")]
        
        # Create search request
        search_request = SemanticSearchRequest(
            query=query,
            limit=limit,
            min_confidence=min_confidence,
            symbol_types=symbol_type_list,
            file_paths=file_path_list,
            include_embeddings=include_embeddings
        )
        
        # TODO: Implement actual semantic search logic
        # For now, return mock response
        mock_results = []
        
        # Create response with constitutional compliance
        response_data = {
            "query": query,
            "results": mock_results,
            "total_results": len(mock_results),
            "search_time_ms": round((time.time() - start_time) * 1000, 2),
            "cache_hit": False
        }
        
        response = SemanticSearchResponse(**ensure_constitutional_compliance(response_data))
        
        # Log API response
        duration_ms = (time.time() - start_time) * 1000
        log_api_response(
            method="GET",
            path="/api/v1/search/semantic",
            status_code=200,
            duration_ms=duration_ms,
            user_id=user_id,
            request_id=request_id,
            results_count=len(mock_results)
        )
        
        return response
        
    except Exception as e:
        logger.error(
            f"Semantic search error: {e}",
            extra={
                "query": query,
                "user_id": user_id,
                "request_id": request_id,
                "constitutional_hash": CONSTITUTIONAL_HASH
            },
            exc_info=True
        )
        
        duration_ms = (time.time() - start_time) * 1000
        log_api_response(
            method="GET",
            path="/api/v1/search/semantic",
            status_code=500,
            duration_ms=duration_ms,
            user_id=user_id,
            request_id=request_id,
            error=str(e)
        )
        
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=ensure_constitutional_compliance({
                "error": "semantic_search_failed",
                "message": "Failed to perform semantic search",
                "request_id": request_id
            })
        )


@api_router.get("/symbols/{symbol_id}", response_model=CodeSymbol, tags=["Symbols"])
async def get_symbol(
    symbol_id: str,
    request: Request,
    current_user: Optional[Dict[str, Any]] = Depends(get_current_user),
    user_id: Optional[str] = Depends(get_user_id),
    request_id: str = Depends(get_request_id)
):
    """
    Get detailed information about a specific code symbol.
    """
    start_time = time.time()
    
    # Log API request
    log_api_request(
        method="GET",
        path=f"/api/v1/symbols/{symbol_id}",
        user_id=user_id,
        request_id=request_id,
        symbol_id=symbol_id
    )
    
    try:
        # TODO: Implement actual symbol lookup logic
        # For now, return 404
        duration_ms = (time.time() - start_time) * 1000
        log_api_response(
            method="GET",
            path=f"/api/v1/symbols/{symbol_id}",
            status_code=404,
            duration_ms=duration_ms,
            user_id=user_id,
            request_id=request_id
        )
        
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=ensure_constitutional_compliance({
                "error": "symbol_not_found",
                "message": f"Symbol with ID {symbol_id} not found",
                "request_id": request_id
            })
        )
        
    except HTTPException:
        raise
    except Exception as e:
        logger.error(
            f"Get symbol error: {e}",
            extra={
                "symbol_id": symbol_id,
                "user_id": user_id,
                "request_id": request_id,
                "constitutional_hash": CONSTITUTIONAL_HASH
            },
            exc_info=True
        )
        
        duration_ms = (time.time() - start_time) * 1000
        log_api_response(
            method="GET",
            path=f"/api/v1/symbols/{symbol_id}",
            status_code=500,
            duration_ms=duration_ms,
            user_id=user_id,
            request_id=request_id,
            error=str(e)
        )
        
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=ensure_constitutional_compliance({
                "error": "symbol_lookup_failed",
                "message": "Failed to retrieve symbol information",
                "request_id": request_id
            })
        )


@api_router.post("/analysis/trigger", response_model=AnalysisResponse, tags=["Analysis"])
async def trigger_analysis(
    analysis_request: AnalysisRequest,
    request: Request,
    current_user: Optional[Dict[str, Any]] = Depends(get_current_user),
    user_id: Optional[str] = Depends(get_user_id),
    request_id: str = Depends(get_request_id)
):
    """
    Trigger code analysis for specific files or the entire codebase.
    """
    start_time = time.time()
    
    # Log API request
    log_api_request(
        method="POST",
        path="/api/v1/analysis/trigger",
        user_id=user_id,
        request_id=request_id,
        analysis_type=analysis_request.analysis_type
    )
    
    try:
        # TODO: Implement actual analysis trigger logic
        # For now, return mock response
        
        mock_job = {
            "id": str(uuid.uuid4()),
            "job_type": analysis_request.analysis_type,
            "status": "queued",
            "file_path": analysis_request.file_paths[0] if analysis_request.file_paths else None,
            "progress_percentage": 0.0,
            "symbols_found": 0,
            "symbols_updated": 0,
            "dependencies_created": 0,
            "embeddings_created": 0,
            "created_at": datetime.now(timezone.utc)
        }
        
        response_data = {
            "job": mock_job,
            "message": f"Analysis job created for {analysis_request.analysis_type}"
        }
        
        response = AnalysisResponse(**ensure_constitutional_compliance(response_data))
        
        # Log API response
        duration_ms = (time.time() - start_time) * 1000
        log_api_response(
            method="POST",
            path="/api/v1/analysis/trigger",
            status_code=202,
            duration_ms=duration_ms,
            user_id=user_id,
            request_id=request_id,
            job_id=mock_job["id"]
        )
        
        return response
        
    except Exception as e:
        logger.error(
            f"Analysis trigger error: {e}",
            extra={
                "analysis_type": analysis_request.analysis_type,
                "user_id": user_id,
                "request_id": request_id,
                "constitutional_hash": CONSTITUTIONAL_HASH
            },
            exc_info=True
        )
        
        duration_ms = (time.time() - start_time) * 1000
        log_api_response(
            method="POST",
            path="/api/v1/analysis/trigger",
            status_code=500,
            duration_ms=duration_ms,
            user_id=user_id,
            request_id=request_id,
            error=str(e)
        )
        
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=ensure_constitutional_compliance({
                "error": "analysis_trigger_failed",
                "message": "Failed to trigger code analysis",
                "request_id": request_id
            })
        )


@api_router.post("/context/enrich", response_model=ContextEnrichmentResponse, tags=["Context"])
async def enrich_context(
    enrichment_request: ContextEnrichmentRequest,
    request: Request,
    current_user: Optional[Dict[str, Any]] = Depends(get_current_user),
    user_id: Optional[str] = Depends(get_user_id),
    request_id: str = Depends(get_request_id)
):
    """
    Enrich code symbols with context from ACGS Context Service.
    """
    start_time = time.time()
    
    # Log API request
    log_api_request(
        method="POST",
        path="/api/v1/context/enrich",
        user_id=user_id,
        request_id=request_id,
        symbol_count=len(enrichment_request.symbol_ids)
    )
    
    try:
        # TODO: Implement actual context enrichment logic
        # For now, return mock response
        
        response_data = {
            "symbol_ids": enrichment_request.symbol_ids,
            "context_links": [],
            "total_links_created": 0,
            "processing_time_ms": round((time.time() - start_time) * 1000, 2)
        }
        
        response = ContextEnrichmentResponse(**ensure_constitutional_compliance(response_data))
        
        # Log API response
        duration_ms = (time.time() - start_time) * 1000
        log_api_response(
            method="POST",
            path="/api/v1/context/enrich",
            status_code=200,
            duration_ms=duration_ms,
            user_id=user_id,
            request_id=request_id,
            links_created=0
        )
        
        return response
        
    except Exception as e:
        logger.error(
            f"Context enrichment error: {e}",
            extra={
                "symbol_count": len(enrichment_request.symbol_ids),
                "user_id": user_id,
                "request_id": request_id,
                "constitutional_hash": CONSTITUTIONAL_HASH
            },
            exc_info=True
        )
        
        duration_ms = (time.time() - start_time) * 1000
        log_api_response(
            method="POST",
            path="/api/v1/context/enrich",
            status_code=500,
            duration_ms=duration_ms,
            user_id=user_id,
            request_id=request_id,
            error=str(e)
        )
        
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=ensure_constitutional_compliance({
                "error": "context_enrichment_failed",
                "message": "Failed to enrich context",
                "request_id": request_id
            })
        )